fix: 修复 double→float 隐式转换编译错误
SubsystemTime.GameTime 返回 double，直接赋值给 float 变量会报
"无法将类型double隐式转换为float"编译错误。

修复2处:
- 第195行 QueuedAtSeconds = s_time.GameTime
  → QueuedAtSeconds = (float)s_time.GameTime
- 第266行 float now = s_time.GameTime;
  → float now = (float)s_time.GameTime;

这两处都在上鞍撤销逻辑(TryConsumePendingRevert/OnEntityRemove)中，
用于暂存/匹配原马待恢复队列的时间戳。

+++ Breeding/SubsystemBreeding.cs
@@ -192,7 +192,7 @@ namespace Game
                             Rotation = body.Rotation,
                             Velocity = body.Velocity,
                             State = state,
-                            QueuedAtSeconds = s_time.GameTime
+                            QueuedAtSeconds = (float)s_time.GameTime
                         });
                         Log.Information($"[Breeding] 暂存禁止交互原马待恢复: template={state.TemplateName}#{entity.Id}, stage={state.GetStageDisplayName()}, pos={body.Position}");
                         s_states.Remove(entity);
@@ -263,7 +263,7 @@ namespace Game
             if (body == null) return false;
             Vector3 pos = body.Position;
 
-            float now = s_time.GameTime;
+            float now = (float)s_time.GameTime;
             for (int i = s_pendingReverts.Count - 1; i >= 0; i--)
             {
                 PendingSaddleRevert r = s_pendingReverts[i];
